+++ setups/apt.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-sudo apt install clang libxkbcommon-x11-dev pkg-config libvulkan-dev libwayland-dev xorg-dev libxcb-shape0-dev libxcb-xfixes0-dev