+++ setups/pacman.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-sudo pacman -S clang vulkan-headers wayland xorg-server-devel xcb-util xcb-util-wm openssl base-devel libxkbcommon-x11