+++ setups/xbps-install.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-sudo xbps-install -S base-devel clang libxkbcommon-devel vulkan-loader wayland-devel